# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: SIGAP API
  version: 1.0.0
  description: Geospatial infrastructure reporting platform API
servers:
  - url: https://sigap.live/api
    description: Production
paths:
...
  /reports:
    get:
      operationId: listReports
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: creator_id
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Paginated reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                  pagination:
                    $ref: "#/components/schemas/Pagination"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createReport
      tags: [reports]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportCreateRequest"
      responses:
        "201":
          description: Report created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  duplicate:
                    type: boolean
        "200":
          description: Report duplicate (idempotency)
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  duplicate:
                    type: boolean
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /reports/{id}:
    get:
      operationId: getReport
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Report detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateReport
      tags: [reports]
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportUpdateRequest"
      responses:
        "200":
          description: Report updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Report"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /public/reports:
    get:
      operationId: listPublicReports
      tags: [public]
      security: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: category_id
          in: query
          schema:
            type: string
            format: uuid
        - name: bbox
          in: query
          schema:
            type: string
        - name: month
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: Public reports list
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      type: object
                  total:
                    type: integer
                  page:
                    type: integer
                  limit:
                    type: integer
    post:
      operationId: createPublicReport
      tags: [public]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublicReportCreateRequest"
      responses:
        "200":
          description: Report created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  duplicate:
                    type: boolean
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Role:
      type: string
      enum: [WARGA, SURVEYOR, VERIFIKATOR, OPERATOR, PETUGAS, PENGAMBIL_KEPUTUSAN, ADMIN_DAERAH, AUDITOR, RT_RW, ADMIN]
    ReportStatus:
      type: string
      enum:
        - draft
        - submitted
        - under_review
        - verified
        - assigned
        - in_progress
        - resolved
        - closed
        - rejected
        - duplicate_merged
        - needs_survey
        - needs_completion
        - out_of_scope
        - merged
        - separated
        - escalated
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        total_pages:
          type: integer
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
        details:
          type: object
          nullable: true

    # Auth
...
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        name:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        deleted_at:
          type: string
          format: date-time
          nullable: true
...
    Category:
      type: object
      properties:
        id:
          type: string
          format: uuid
        slug:
          type: string
        name:
          type: string
        icon:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        parent_id:
          type: string
          format: uuid
          nullable: true
        code:
          type: string
          nullable: true
        short_code:
          type: string
          nullable: true
        color_class:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
...
    Report:
      type: object
      properties:
        id:
          type: string
          format: uuid
        idempotency_key:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        category:
          $ref: "#/components/schemas/Category"
          nullable: true
        description:
          type: string
        geom:
          type: object
          nullable: true
        lat:
          type: number
        lng:
          type: number
        photo_urls:
          type: array
          items:
            type: string
        exif_data:
          type: object
          nullable: true
        device_id:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ReportStatus"
        severity:
          type: integer
          minimum: 1
          maximum: 5
          nullable: true
        assigned_to:
          type: string
          format: uuid
          nullable: true
        assignee:
          $ref: "#/components/schemas/User"
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ReportCreateRequest:
      type: object
      required: [idempotency_key, category_id, description, lat, lng]
      properties:
        idempotency_key:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        description:
          type: string
          minLength: 1
          maxLength: 2000
        lat:
          type: number
        lng:
          type: number
        device_id:
          type: string
          format: uuid
        photo_urls:
          type: array
          items:
            type: string
            format: uri
        reported_at:
          type: string
          format: date-time
        title:
          type: string
          maxLength: 255
        population_affected:
          type: integer
          minimum: 0
        vulnerability_index:
          type: number
          minimum: 0
          maximum: 1
        consent:
          type: boolean
    ReportUpdateRequest:
      type: object
      properties:
        status:
          $ref: "#/components/schemas/ReportStatus"
        description:
          type: string
          minLength: 1
          maxLength: 2000
        priority:
          type: integer
          minimum: 0
          maximum: 100
        assigned_to:
          type: string
          format: uuid
          nullable: true
    PublicReportCreateRequest:
      type: object
      required: [idempotency_key, category_id, description, lat, lng, device_id]
      properties:
        idempotency_key:
          type: string
          format: uuid
        category_id:
          type: string
          format: uuid
        description:
          type: string
          minLength: 10
          maxLength: 500
        lat:
          type: number
        lng:
          type: number
        device_id:
          type: string
          format: uuid
        photo_urls:
          type: array
          items:
            type: string
            format: uri
        reported_at:
          type: string
          format: date-time
        title:
          type: string
          maxLength: 255
        population_affected:
          type: integer
          minimum: 0
        vulnerability_index:
          type: number
          minimum: 0
          maximum: 1
        consent:
          type: boolean
...